+++ index.js
@@ -1,5 +1,5 @@
 const RedisFeatureStore = require('./redis_feature_store');
-const RedisBigSegmentStore = require('./redis_big_segment_store');
+const { RedisBigSegmentStore } = require('./redis_big_segment_store');
 
 module.exports = {
   RedisFeatureStore,
